% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicas.R
\name{correlacion}
\alias{correlacion}
\title{Correlación rápida}
\usage{
correlacion(x, y, metodo = c("pearson", "spearman", "kendall"))
}
\arguments{
\item{x, y}{Vectores numéricos.}

\item{metodo}{"pearson", "spearman" o "kendall".}
}
\value{
Correlación entre \code{x} y \code{y}.
}
\description{
Correlación rápida
}
\examples{
correlacion(1:10, 1:10)
}
