% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicas.R
\name{cuantiles}
\alias{cuantiles}
\title{Cuantiles}
\usage{
cuantiles(x, probs = c(0.25, 0.5, 0.75))
}
\arguments{
\item{x}{Vector numérico.}

\item{probs}{Probabilidades (0-1). Por defecto 0.25, 0.5 y 0.75.}
}
\value{
Vector con cuantiles.
}
\description{
Cuantiles
}
\examples{
cuantiles(1:9)
}
