% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicas.R
\name{desviacion}
\alias{desviacion}
\title{Desviación típica con NA ignorados}
\usage{
desviacion(x)
}
\arguments{
\item{x}{Vector numérico.}
}
\value{
Desviación estándar muestral de \code{x}.
}
\description{
Desviación típica con NA ignorados
}
\examples{
desviacion(1:5)
}
