% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicas.R
\name{estandarizar}
\alias{estandarizar}
\title{Estandarización (z-score)}
\usage{
estandarizar(x)
}
\arguments{
\item{x}{Vector numérico.}
}
\value{
(x - media)/sd con NA ignorados.
}
\description{
Estandarización (z-score)
}
\examples{
estandarizar(1:10)
}
