% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicas.R
\name{grafica}
\alias{grafica}
\title{Gráfica rápida (dispersión o línea / histograma)}
\usage{
grafica(x, y = NULL, tipo = "p")
}
\arguments{
\item{x}{Vector numérico.}

\item{y}{Opcional: si se pasa, hace plot(x,y). Si no, hist(x).}

\item{tipo}{Tipo de gráfico base R para \code{plot} ("p" puntos, "l" líneas, etc.).}
}
\description{
Gráfica rápida (dispersión o línea / histograma)
}
\examples{
grafica(1:10, (1:10)^2)
grafica(rnorm(100))
}
