% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicas.R
\name{grafica_residuos}
\alias{grafica_residuos}
\title{Diagnóstico sencillo de residuos de un modelo lineal}
\usage{
grafica_residuos(modelo)
}
\arguments{
\item{modelo}{Objeto \code{lm}.}
}
\description{
Diagnóstico sencillo de residuos de un modelo lineal
}
\examples{
m <- lm(mpg ~ wt, data = mtcars); grafica_residuos(m)
}
