% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicas.R
\name{media}
\alias{media}
\title{Media (promedio) con NA ignorados}
\usage{
media(x)
}
\arguments{
\item{x}{Vector numérico.}
}
\value{
Valor medio de \code{x}.
}
\description{
Media (promedio) con NA ignorados
}
\examples{
media(1:5)
}
