% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misalumnos-package.R
\docType{package}
\name{misalumnos-package}
\alias{misalumnos}
\alias{misalumnos-package}
\title{misalumnos: utilidades y ejemplos para Métodos Estadísticos}
\description{
Este paquete incluye funciones básicas (por ejemplo, normalización y distribuciones discretas)
y conjuntos de datos de ejemplo usados en las prácticas de Métodos Estadísticos.
}
\author{
\strong{Maintainer}: Tu Nombre \email{tu@email.com}

}
\keyword{internal}
