% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicas.R
\name{normalizar}
\alias{normalizar}
\title{Normalización 0,1}
\usage{
normalizar(x)
}
\arguments{
\item{x}{Vector numérico.}
}
\value{
Vector reescalado a 0,1. Si min==max devuelve ceros.
}
\description{
Normalización 0,1
}
\examples{
normalizar(1:10)
}
