% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicas.R
\name{regresion_lineal}
\alias{regresion_lineal}
\title{Regresión lineal simple}
\usage{
regresion_lineal(x, y)
}
\arguments{
\item{x}{Predictor numérico.}

\item{y}{Respuesta numérica.}
}
\value{
Objeto \code{lm}.
}
\description{
Regresión lineal simple
}
\examples{
set.seed(1); x <- 1:10; y <- x + rnorm(10); m <- regresion_lineal(x,y); summary(m)
}
