% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicas.R
\name{varianza}
\alias{varianza}
\title{Varianza con NA ignorados}
\usage{
varianza(x)
}
\arguments{
\item{x}{Vector numérico.}
}
\value{
Varianza muestral de \code{x}.
}
\description{
Varianza con NA ignorados
}
\examples{
varianza(1:5)
}
